Changes made to official firmware for DEMIURGE project

+++ IREmission.h
@@ -6,9 +6,11 @@
 #define NUM_RECEPTOR_SENSORS 12
 
 /* Store and Write the IR data sent to all the sensors. The same for all*/
-void WriteAllIrData ( unsigned int data);
+void WriteAllIrData ( unsigned long data);
+//void WriteAllIrData ( unsigned int data);
 /* Store IR data on each specific channel */
-void StoreIrData ( int channel, unsigned int data);
+void StoreIrData ( int channel, unsigned long data);
+//void StoreIrData ( int channel, unsigned int data);
 /*Send the loaded values */
 void WriteIrData ( void );
