+++ registers.h
@@ -14,18 +14,21 @@ typedef struct {
 	int received;
 	int alpha;
 	unsigned int distance;
-	unsigned int data;
+	//unsigned int data;
+	unsigned long data;
 	char enable;
 } dataRegisterReceived;
 
 typedef struct {
 	int received;
 	unsigned int alpha;
-	unsigned int data;
+	unsigned long data;
+	//unsigned int data;
 } modelledDataRegisterReceived;
 
 typedef struct {
-	unsigned int data;
+	//unsigned int data;
+	unsigned long data;
 	float angle;
 	unsigned int range;
 	unsigned int max_peak1;
@@ -37,12 +40,14 @@ typedef struct {
 #endif
 
 void setDataReceived (dataRegisterReceived data);
-void setAllDataReceived (dataRegisterReceived *Alldata, unsigned int data);
+//void setAllDataReceived (dataRegisterReceived *Alldata, unsigned int data);
+void setAllDataReceived (dataRegisterReceived *Alldata, unsigned long data);
 void setModelledDataReceived (modelledDataRegisterReceived data);
 
 void getDataReceived (dataRegisterReceived *data);
 //void getAllDataReceived (dataRegisterReceived *data);
-unsigned char getAllDataReceived ( dataRegisterReceived *Alldata, unsigned int *data);
+//unsigned char getAllDataReceived ( dataRegisterReceived *Alldata, unsigned int *data);
+unsigned char getAllDataReceived ( dataRegisterReceived *Alldata, unsigned long *data);
 void getModelledDataReceived (modelledDataRegisterReceived *data);
 
 void resetDataReceived ( void );
